Move app_utils and email_utils into skyportal/utils, and drop the unused server.protocol config value.
