+++ config.yaml.defaults
@@ -712,7 +712,6 @@ server:
   # - Authorized redirect URLs: http://localhost:5000/complete/google-oauth2/
   #
   # You need to have Google+ API enabled; it takes a few minutes to activate.
-  protocol: http
   host: localhost
   port: 5000  # This is the public-facing port
