+++ services/reminders/reminders.py
@@ -6,8 +6,6 @@ from baselayer.app.env import load_env
 from baselayer.app.flow import Flow
 from baselayer.app.models import User, init_db
 from baselayer.log import make_log
-from skyportal.app_utils import get_app_base_url
-from skyportal.email_utils import send_email
 from skyportal.models import (
     DBSession,
     Reminder,
@@ -18,6 +16,8 @@ from skyportal.models import (
 )
 from skyportal.models.gcn import GcnEvent
 from skyportal.models.shift import Shift
+from skyportal.utils.app import get_app_base_url
+from skyportal.utils.email import send_email
 from skyportal.utils.naive_datetime import utcnow_naive
 from skyportal.utils.services import check_loaded
