+++ skyportal/handlers/api/analysis.py
@@ -27,7 +27,6 @@ from baselayer.app.flow import Flow
 from baselayer.app.model_util import recursive_to_dict
 from baselayer.log import make_log
 
-from ...app_utils import get_app_base_url
 from ...enum_types import (
     ANALYSIS_INPUT_TYPES,
     ANALYSIS_TYPES,
@@ -49,6 +48,7 @@ from ...models import (
     User,
     UserNotification,
 )
+from ...utils.app import get_app_base_url
 from ...utils.extinction import calculate_extinction
 from ...utils.naive_datetime import utcnow_naive
 from ...utils.parse import str_to_bool
