+++ skyportal/models/gcn.py
@@ -48,6 +48,7 @@ from baselayer.app.models import (
     safe_aliased,
 )
 
+from ..utils.app import get_app_base_url
 from ..utils.cache import Cache, cache_folder, dict_to_bytes
 from .allocation import Allocation, AllocationUser
 from .group import accessible_by_group_members, accessible_by_groups_members
@@ -55,9 +56,7 @@ from .localization import Localization
 
 env, cfg = load_env()
 
-host = f"{cfg['server.protocol']}://{cfg['server.host']}" + (
-    f":{cfg['server.port']}" if cfg["server.port"] not in [80, 443] else ""
-)
+host = get_app_base_url()
 
 cache_dir = f"{cache_folder}/public_pages/reports"
 cache = Cache(
