+++ skyportal/utils/app.py
