+++ skyportal/utils/notifications.py
@@ -10,11 +10,11 @@ from sqlalchemy.orm import selectinload
 
 from baselayer.app.env import load_env
 from baselayer.log import make_log
-from skyportal.app_utils import get_app_base_url
-from skyportal.email_utils import send_email
 from skyportal.models import GcnEvent
 from skyportal.models.gcn import SOURCE_RADIUS_THRESHOLD
+from skyportal.utils.app import get_app_base_url
 from skyportal.utils.calculations import deg2dms, deg2hms, radec2lb
+from skyportal.utils.email import send_email
 
 from .naive_datetime import utcnow_naive
